requirements.txt

cffi==1.17.1
configobj==5.0.9
cryptography==37.0.4
future==1.0.0
portalocker==2.0.0
psutil==7.0.0
pycklink==0.1.1
pycparser==2.22
pylink-square==0.5.0
pyserial==3.5
six==1.17.0
toml==0.10.0
setuptools==80.9.0
